refactor(db): split models.py into _models/ package
