+++ backend/src/ledger_sync/db/_models/_constants.py
@@ -1,0 +1,5 @@
+"""Shared constants used across model modules."""
+
+# Foreign key reference constant for user relationships
+USER_FK = "users.id"
+CASCADE_ALL_DELETE_ORPHAN = "all, delete-orphan"
